+++ __tests__/data-tests.js
@@ -16,7 +16,6 @@ test('Query without id should be filtered by limit', async () => {
     let messages = (await client.query({
         query: gql`query { messages(limit: 1){value created_at created_lt} }`,
     }));
-console.log(messages)
     expect(messages.data.messages.length).toEqual(1);
     server.stop();
 });
@@ -45,7 +44,7 @@ test('Data Broker', async () => {
         logs: new QLogs(),
         data: createTestData({
             mutable: mut,
-            immutable: new QDataCombiner([hot, new QDataPrecachedCombiner(cache, cold)]),
+            immutable: new QDataCombiner([hot, new QDataPrecachedCombiner(cache, cold, testConfig.networkName, testConfig.cacheKeyPrefix)]),
         }),
     });
     await server.start();
@@ -90,4 +89,3 @@ test('Data Broker', async () => {
 
     server.stop();
 });
-