+++ src/server/server.js
@@ -103,7 +103,7 @@ class MemStats {
     }
 }
 
-export function createProviders(configName: string, logs: QLogs, config: QDataProvidersConfig): QDataProviders {
+export function createProviders(configName: string, logs: QLogs, config: QDataProvidersConfig, networkName: string, cacheKeyPrefix: string): QDataProviders {
     const newArangoProvider = (dbName: string, config: QArangoConfig): ArangoProvider => (
         new ArangoProvider(logs.create(`${configName}_${dbName}`), config)
     );
@@ -112,6 +112,8 @@ export function createProviders(configName: string, logs: QLogs, config: QDataPr
     const cold = new QDataPrecachedCombiner(
         isCacheEnabled(config.cache) ? new MemjsDataCache(logs.create(`${configName}_cache`), config.cache) : missingDataCache,
         config.cold.map(x => newArangoProvider('cold', x)),
+        networkName,
+        cacheKeyPrefix,
     );
     const immutable = new QDataCombiner([hot, cold]);
     return {
@@ -152,8 +154,8 @@ export default class TONQServer {
             auth: this.auth,
             tracer: this.tracer,
             stats: this.stats,
-            providers: createProviders('fast', this.logs, this.config.data),
-            slowQueriesProviders: createProviders('slow', this.logs, this.config.slowQueriesData),
+            providers: createProviders('fast', this.logs, this.config.data, this.config.networkName, this.config.cacheKeyPrefix),
+            slowQueriesProviders: createProviders('slow', this.logs, this.config.slowQueriesData, this.config.networkName, this.config.cacheKeyPrefix),
             isTests: false,
         });
         this.memStats = new MemStats(this.stats);